MAVProxy/pymavlink/requirements.txt
lxml>=3.6.0
future>=0.15.2
wheel>=0.37.1
setuptools>=42
fastcrc

# crypto
pycryptodome
twofish
psutil

# dev dependencies:
pytest<=7.4.4
syrupy; python_version>='3.6'
